third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/calc-size-height-interpolation.tentative-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_equals: expected "75px " but got "200px "
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0) should be [100px]
  assert_equals: expected "100px " but got "200px "
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_equals: expected "125px " but got "200px "
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_equals: expected "150px " but got "200px "
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_equals: expected "175px " but got "200px "
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_equals: expected "225px " but got "200px "
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_equals: expected "75px " but got "200px "
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0) should be [100px]
  assert_equals: expected "100px " but got "200px "
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_equals: expected "125px " but got "200px "
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_equals: expected "150px " but got "200px "
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_equals: expected "175px " but got "200px "
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_equals: expected "225px " but got "200px "
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_equals: expected "150px " but got "200px "
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_equals: expected "175px " but got "200px "
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_equals: expected "225px " but got "200px "
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_equals: expected "150px " but got "200px "
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_equals: expected "175px " but got "200px "
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_equals: expected "225px " but got "200px "
Harness: the test ran to completion.
